Provide composition for running ccbv
Provides a docker composition for running ccbv locally. This was done
a bit slapdash. The settings for this application aren't fully
customizable via environment variables. But where possible the
customizations were made to make this work within a container.

It's clear the application was developed with a checkout and run
mentality. There is nothing wrong with this, it just makes
containerizing the app for development slightly more challenging.